--- src/cosmos-sdk/coin/deccoin.ts
@@ -1,4 +1,0 @@
-export interface DecCoin {
-  denom: string;
-  amount: bigint;
-}
