+++ src/types/cosmos-sdk/coin.ts
