+++ src/types/cosmos-sdk/deccoin.ts
@@ -1,0 +1,6 @@
+import { Dec } from "./decimal";
+
+export interface DecCoin {
+  denom: string;
+  amount: Dec;
+}
