+++ src/types/cosmos-sdk/decimal.ts
@@ -1,0 +1,1 @@
+export type Dec = bigint;
