+++ src/types/tendermint/priv-key.ts
@@ -1,8 +1,4 @@
-export interface PubKey {
-  toBuffer(): Buffer;
-  toBase64(): string;
-  verify(message: string, signature: Buffer): boolean;
-}
+import { PubKey } from "./pub-key";
 
 export interface PrivKey {
   getPubKey(): PubKey;
