+++ src/types/tendermint/pub-key.ts
@@ -1,0 +1,5 @@
+export interface PubKey {
+  toBuffer(): Buffer;
+  toBase64(): string;
+  verify(message: string, signature: Buffer): boolean;
+}
