+++ src/x/auth/auth.ts
@@ -1,9 +1,9 @@
-import { BroadcastReq } from "./broadcast-req";
-import { BaseAccount } from "./account";
+import { BroadcastReq } from "./types/broadcast-req";
+import { BaseAccount } from "./types/account";
 import { CosmosSdkHost } from "../../common/cosmos-sdk-host";
-import { TxResponse, SearchTxsResult } from "../../cosmos-sdk/result";
-import { StdTx } from "./stdtx";
-import { EncodeResp } from "./encode-resp";
+import { TxResponse, SearchTxsResult } from "../../types/cosmos-sdk/result";
+import { StdTx } from "./types/stdtx";
+import { EncodeResp } from "./types/encode-resp";
 
 /**
  * Cosmos SDKにおけるx/authのRest APIをまとめたモジュール。
