+++ src/x/bank/types/send-req.ts
@@ -1,0 +1,7 @@
+import { BaseReq } from "../../../types/cosmos-sdk/rest";
+import { Coin } from "../../../types/cosmos-sdk/coin";
+
+export interface SendReq {
+  base_req: BaseReq;
+  amount: Coin[];
+}
